do collisions and ability to add multiple following enemies

+++ src/utility/Position.java
@@ -4,8 +4,8 @@ package utility;
  * Created by student on 5/4/17.
  */
 public class Position {
-    int x, y;
-    public Position(int x, int y) {
+    float x, y;
+    public Position(float x, float y) {
         this.x = x;
         this.y = y;
     }
@@ -15,10 +15,10 @@ public class Position {
         this.y = position.y;
     }
 
-    public void moveLeft(int off) { this.x -= off; }
-    public void moveRight(int off) { this.x += off; }
-    public void moveUp(int off) { this.y -= off; }
-    public void moveDown(int off) { this.y += off; }
-    public int getYPosition() { return this.y; }
-    public int getXPosition() { return this.x; }
+    public void moveLeft(float off) { this.x -= off; }
+    public void moveRight(float off) { this.x += off; }
+    public void moveUp(float off) { this.y -= off; }
+    public void moveDown(float off) { this.y += off; }
+    public float getYPosition() { return this.y; }
+    public float getXPosition() { return this.x; }
 }
